testsuite: cover job behavior with slow scheduler
Problem: there are no tests that show jobs can make progress
they are not blocked on an allocation, but the scheduler is slow.

Add a test that uses the new module-debug "defer" function
to simulate a slow scheduler and make sure that jobs can
get through CLEANUP when the scheduler is out to lunch,
and that the system recovers and can run jobs aftwards.