broker: honor control disconnect during hello
Problem: if the parent broker reboots after hello was sent
but before a response is received, the child broker is stuck
unable to connect.

As noted in #5881, we see a continuous stream of log messages on
the client:

   DROP upstream control topic - :
     message received before hello handshake completed

The messages may get started because
1. child sends hello
2. parent reboots and misses the hello
3. child starts sending periodic heartbeat control message to
   the parent after 5s regardless of the hello handshake status
   (this is by design - see below)
4. parent expects to get a hello request before any other message,
   so it sends a disconnect control message
5. child logs the above message (which goes to the parent)
   but doesn't disconnect
6. parent recieves log message, goto 4

Note that although the broker's downstream 0MQ ROUTER socket exposes
TCP disconnects as you might expect for a regular socket, the upstream
0MQ DEALER socket does not.  This is why the child does not detect the
parent restart after sending hello - it has seamlessly reconnected to
the new parent but the hello request that was likely safely delivered to
the old parent before it died and is gone.

The design to work around this passive reconnect behavior is to start
the periodic heartbeat control messages early and force something to happen.
The parent does the right thing by sending the disconnect in response to
the heartbeat.  Unfortunatley, the child ignores the disconnect and a
fast but unentertaining game of pingpong ensues.

To break the cycle, handle the control disconnect message in the child
as originally intended.  That will cause the broker to be restarted by
systemd and introductions can be restarted from scratch.

This first noted in #5881.