systemd: change RestartSec=5s to 30s
Problem: when nodes of a system instance are forcibly disconnected,
they can reconnect fairly quickly because we allow systemd to
restart them in 5s.

While Flux is shutting down, hello requests are now immediately sent
a control disconnect message, but if the fanout is large and the
shutdown is slow, the whack-a-mole overhead may be non-negligible.

Raise the systemd unit file RestartSec value from 5s to 30s.