broker: improve LOST error message
Problem: many "transitioning to LOST due to EHOSTUNREACH error on send"
messages were logged during shutdown of a large instance.

This is still not well understood but we can perhaps get a little more
information for next time.

Add the previous state and the time the peer has spent in that
state (in whole seconds).

Hopefully will help with #5881 if it occurs again.