libjob: add flux_job_waitstatus_to_exit_code(3)
Problem: There are at least two places where a wait status from a
job is converted to an exit code with a possible error message on
non-zero exit status. This could result in differences in behavior
over time if one location is checked and another is not.

Add flux_job_waitstatus_to_exit_code(3) which can translate a job
waitstatus to an exit code using the following rules:

 If the job shell was killed by a signal, then return 128+signo,
 else exit with the result of WEXITSTATUS (waitstatus)

Also, if the job exits with nonzero exit code, then return an error
string in a flux_error_t parameter, noting the non-zero exit code,
or if the job shell or tasks were apparently killed by a signal.