terminus/test: fix potential segfault in test_terminus.t
Problem: The terminus server function in the terminus unit tests
calls setenv(3), but this function is run in a thread and setenv(3)
is not thread safe. This could result in rare segfaults if reading
the environment races with setting it during the test.

Move the setenv to main(), there's no reason to be setting the
environment variable specifically in terminus_server() anyway,
since this is a unit test, all invocations of the server should have
FLUX_TERMINUS_TEST_SERVER set.

Fixes #3120