t: update command usage with -L (--color) option
Problem: In the near future the -L short option for --color can
take an optional parameter.  The usage in a test:

flux job eventlog -HLp output $(flux job last)

will break as a result since the -L option now believes "p" is the argument
to -L.

Update the usage to

flux job eventlog -HL -p output $(flux job last)

so it is clear -L is not being given an argument.