flux-exec: set up dbus environment for systemctl
Problem: it is tricky to run systemd commands using flux-exec(1)
on a system instance configured with exec.service=sdexec
because the user systemd environment is not set up.

If not already set, set DBUS_SESSION_BUS_ADDRESS so that systemd
--user commands access the dbus instance running as the flux user.
Thus commands like this can work (after sudo to root or flux):

  flux exec -r 3 systemctl --user list-units --type=service
  flux exec -r 3 systemctl --user status imp-shell-3-ftSp9MrEVNw.service

Fixes #5901.