flux-perilog-run: raise job exception for offline rakns
Problem: When a prolog fails due to offline ranks, a generic exception
like "prolog exited with exit code=1" is raised after the prolog
exits, but this gives the user and admins no details about why the
prolog failed.

Add a nonfatal exception from the job prolog when there are offline
ranks detected so that a specific reason is associated with the
job eventlog.