flux-config: add get --config-path=PATH option
Problem: there is no easy way to parse-check a Flux configuration
without loading it into the broker.

Add flux config get --config-path=PATH which tells flux config get
to load the configuration from a file rather than from the live broker.
Since flux_conf_parse() is used internally, PATH can be a TOML file,
a JSON file, or a directory of TOML files.  In addition, for convenience,
the special values of "system", "security", and "imp" are expanded to
$sysconfdir/flux/PATH/conf.d.

Fixes #5898