testsuite: test handling of offline ranks from job prolog
Problem: In t2274-manager-perilog.t, handling of offline ranks by
flux-perilog-run.py is only tested after a job has run by manually
invoking the command. However, this doesn't ensure that the script
raises a nonfatal exception of the job instead of silently failing.

Use the new offline.so jobtap plugin to force rank 3 offline instead
of disconnecting the rank outside of any job. Since the offline.so
plugin is loaded before the perilog.so plugin, this simulates a rank
going offline between the scheduler assigning resources and the prolog.

Ensure that a nonfatal exception is raised on the job.