flux-job: note stopped queues in attach status line
Problem: When the target queue for a job is stopped, it may be
mystifying why a job is not running where there appear to be plenty
of resources available. Currently, the `flux job attach` status line
indicates only "<jobid> waiting for resources" during this condition,
but the tool could easily check for a stopped queue and amend the
status message with a note that could clear up this confusion.

Check for stopped queues in `flux job attach` and amend the "waiting
for resources" status line with "<name> queue stopped" if the target
queue for the job is currently stopped.

Fixes #5896