resource: add EVENT_NO_COMMIT flag
Problem: the KVS resource.eventlog can grow quite large on a
big/busy system, but some events do not need to be persistent.

Add an EVENT_NO_COMMIT flag similar to the one in used for the
job-manager journal.  Events posted with this flag will be seen in
the internal callback but will not go to the KVS.

Define the flag.
Add a flags parameter to reslog_post_pack().
Update callers to pass in 0 for the flags parameter.