resource: don't post resource-init event
Problem: the resource-init event complicates drain/undrain replay
for very little gain.

All the context keys of resource-init except "drain" are ignored on
replay, so those are purely informational.  The resource-init event
itself marks resource module reload, but is probably not too useful
given that it is always followed by a resource-define event.

The drain key summarizes the drain state at that point based on a replay
of the events that came before.  On restart, all drain/undrain events
before the most recent resource-init event are ignored.  The main
purpose of resource-init seems to be an optimization to avoid replaying
a large number of drain events that could accumulate over time, without
addressing the unchecked eventlog growth in the KVS.

Stop posting the resource-init event, and ignore it if present.  All
drain/undrain events in the resource.eventlog are replayed on restart.
The unchecked eventlog growth will be dealt with separately.

This change should allow flux-core to be upgraded or downgraded without
issue.  When upgrading, any existing resource-init events are simply
ignored.  When downgrading, missing resource-init events should not be
a problem as the drain code will replay drain/undrain events since the
last resource-init event.