resource: include nodelist in drain events
Problem: when a cluster is reconfigured so that Flux ranks map to
different hostnames, or if a cluster size is reduced, Flux may
drain the wrong nodes or fail on restart.

As proposed in #4791, include a nodelist in the drain/undrain events
so it becomes possible to make them resilient to such changes.

The additional key is ignored by prior flux-core releases, so
downgrading back to an older release should not cause problems.