resource: handle drain rank:host remapping
Problem: when drain/undrain events are replayed on instance restart,
if the rank:host mapping has changed, the wrong nodes could be
affected or the replay could fail, e.g. if the instance size shrank.

If a nodelist key is present in a drain/undrain event, use it to
update the target ranks.  If the a hostname in the nodelist is unknown
or if a rank is out of range, simply ignore it rather than aborting
replay.

A missing nodelist just causes the reranking to be skipped so after this
change, old eventlogs can still be ingested.

Note that the broker 'hostlist' attribute is used to query the current
mapping rather than R because R may not yet be in inventory when the
eventlog is replayed.

Fixes #4791