perilog: do not set working directory on subprocess
Problem: The job-manager perilog plugin manually sets the cwd of
any subprocesses run for a prolog or epilog, but this is unnecessary
and furthermore forces a fallback to the fork/exec method instead of
posix_spawn(3) which is more performant.

Remove the use of `flux_cmd_setcwd()` in the perilog plugin.