libflux: fix possible use-after-free in chained futures
Problem: If the next future in a chained future is destroyed before
the previous future, then a use-after-free can occur if the previous
future is eventually fulfilled, since chained_continuation() and/or
flux_future_continue(3) reference the now destroyed cf->next. This
condition can occur, for example, if a timeout is specified in a call
to flux_future_then(3) on cf->next, but can also occur in cases where
cf->next is destroyed early for other reasons.

Additionally, a reasonable expectation is that when next is destroyed,
this prevents prev (and any other previous futures) callbacks from being
called, either by destroying them or by other means.

Arrange to notify the chained_future object when cf->next is destroyed
by using an aux_item destructor. Destroy cf->prev if it is non-NULL.

Ensure cf->prev is not subject to double-free by nullifying it when
cf->prev is destroyed. This allows cf->prev and cf->next to be destroyed
in any order safely.

Since cf is associated in the aux_item list of cf->prev, and it is now
referenced by cf->next as well, add a refcount to this object and take
a reference for both cf->prev and cf->next. When both references are
released the object will be destroyed.

Fixes #5923