broker: post goodbye event from state machine
Problem: it is rather confusing to trace how the broker transitions
out of GOODBYE state on all ranks, since the "goodbye" event is
posted only in shutdown.c, which doesn't do much on followers.

Define an action_goodbye() callback, following the pattern used in
other states, and have it post the "goodbye" event for followers.

This will enable other things to be done in the action callback too.