broker: add goodbye handshake
Problem: the shutdown handshake is insufficient to prevent the
parent from accessing a child peer's socket after it has disconnected
and logging errors about unexpected disconnection.

Currently a child node immediately exits GOODBYE state, and tears
down the overlay network, which sends a control status offline message
and disconnects the socket.  The parent processes the control status
message asynchronously (whenever it emerges from the message queue).
In the mean time, the parent could send messages (for example routine
event broadcasts) to the child's socket.  If the send fails with
EHOSTUNREACH before the peer is marked offline, the unexpected
disconnection is logged at LOG_ERR.

Add a overlay.goodbye RPC which the child sends to its parent when it
enters GOODBYE state.  The child remains in GOODBYE state until a
response is received (or a timeout), then it proceeds to EXIT state and
the disconnects as before.

While waiting for the response, further messages to the parent,
like heartbeats or log messages, are suppressed.  This is to avoid
triggering a disconnect control message from the parent if the
follower has been marked OFFLINE before the message is processed,
as it surely will be given message order guarantees.  The disconnect
control messsage causes immediate disconnection - not logged on the
parent, but LOG_CRIT on the child.

Fixes #5881