job-exec: support config stats callbacks
Problem: In the near future we'd like the job-exec module to return
some module information via a stats callback, but no callback exists.
In addition, we'd like stats to be possible for each exec implementation.

Add the initial infrastructure for a stats message handler and exec
implementation callback.  For the time being, no stats are actually
reported.