shell: add hwloc.restrict option
Problem: The HWLOC_XMLFILE provided by the hwloc.xmlfile job shell
option is not restricted to the resources available to the job,
but in some circumstances this would be useful.

Add a hwloc.restrict option that, when used with hwloc.xmlfile,
restricts the HWLOC_XMLFILE to the current process binding.