libsubprocess: server: do not log when writing to process fails
Problem: There's an inherent race between processes being killed or
exiting early and stdin data for those procesess which results in a
lot of errors such as

 Error writing N bytes to subprocess pid PID: unknown pid

This is common enough that this just becomes log noise. Since there
isn't much to be done about it and it is an expected condition,
just remove the log message and silently drop the data.