libsubprocess: add notes for EOF when reading line
Problem: When reading a line via flux_subprocess_read_line() or
flux_subprocess_read_trimmed_line(), a length of 0 cannot be assumed
to mean EOF on a stream.  It's possible there is buffered data
available, but just not a line's worth of data.

Add a comment indicating this fact and that
flux_subprocess_read_stream_closed() should be used to determine
if the stream is really closed.