libsubprocess: refactor read API
Problem: The flux_subprocess_read() family of functions
return a pointer to a buffer on success and length of buffer in an
optional parameter.

This API style isn't the most intuitive as it is quite different
than the read(2) system call.  Most notably when EOF
is reached on a read stream you have to check if the buffer pointer
is non-NULL and the length returned from an optional parameter
is 0.

Solution: Update the read family of functions to return the
length from the function and the buffer pointer in an optional parameter.

Update all callers accordingly.

Fixes #5953