rename variables in calls to libsubprocess
Problem: In the near future the API for several libsubprocess functions
will change.  As a result, some variable names that were previously
used may not be the most useful/meaningful afterwards.

In calls to flux_subprocess_read() and similar functions, use the
variable "buf" instead of the generic "ptr" when referring the
buffer of data returned.  Use the variable "len" instead of "lenp"
for the data length.