libsubprocess: differentiate test descriptions
Problem: A number of tests in test/remote.c output test descriptions
based on the command executed in the subprocess.  However, multiple
tests are run under "/bin/sh", making those test descriptions duplicated
across multiple tests.  It'd be nice if these test descriptions could
be differentiated.

To differentiate the tests, instead specify a "prefix" parameter that
will be output before a collection of tests.  The prefix can be set to
something unique to differentiate tests.