job-exec: use libsubprocess local unbuf flags
Problem: Tests show that when launching many remote subprocesses
through job-exec, there can be performance slowdown due to libsubprocess.

Use the new libsubprocess FLUX_SUBPROCESS_FLAGS_LOCAL_UNBUF to improve
performance.  This flag will skip the buffering of data before being
sent to the caller.  In additiona, it will not bother to NUL terminate
the data, which job-exec does not need since it is immediately sent to
the eventlog.