alloc-bypass: post free event outside of plugin
Problem: if flux is restarted with a running alloc-bypass job,
a free event is not posted when the job finishes, causing it to
hang.

This is because the alloc-bypass aux items are not attached to the
job when it restarts because the job.validate callback is not called.

event_job_action() already has a conditional in FLUX_JOB_STATE_CLEANUP
to handle job->alloc_bypass, but it skips both the sched.free request
and posting the free event.  Alter it so that it posts the free event.

This is simplier than registering a callback for job.cleanup in the plugin
to recover the alloc-bypass state, and it covers the case where the
plugin is not loaded on instance restart.

Fixes #5981