flux-dump: add --ignore-failed-read option
Problem: it may be helpful to ignore content load errors when
creating a dump for garbage collection/recovery purposes.

A flux dump following a disk write failure could fail if any KVS
KVS metadata points to nonexistent blobrefs.  Skipping them and
continuing with the dump would allow the bad metadata to be automatically
purged when flux restarts from the dump.

Add a --ignore-failed-read option to flux-dump(1) similar to the tar(1)
option.  Keys referencing bad blobrefs are omitted and a message is emitted
to stderr, but flux-dump(1) still returns success.

Fixes #5984