rc3: use flux dump --ignore-failed-read
Problem: when the KVS contains dangling blobrefs, the KVS entries
must be manually removed.

Since we now dump the system instance at every shutdown, call it
with the --ignore-failed-read option so that any bad entries are
omitted from the dump.  Then a flux restart should take care of
any problems.