libsubprocess: handle remote protocol errors
Problem: the remote subprocess client assumes a subprocess has
exited if the server returns ENODATA.  This is the case even if
a finished response was not sent, as required by RFC 42.  This
should be handled as a protocol error.

Solution: If the status stream ends early, do not assume the remote
process has exited.  Instead, handle this as an EPROTO error and
fail the subprocess.

Fixes #5956