shell: output: log about output truncation only once
Problem: The job shell currently logs multiple warnings when there
is output truncation. This is excessive and also defeats the purpose
of truncating amount of data going to the KVS.

Move the logging of this warning to when the output plugin is shutting
down so the message is only printed once per stream.

Fixes #5992