libsubprocess: fail flux_subprocess_kill w/ ESRCH
Problem: when flux_subprocess_kill() is called on a local
process but there is nothing to kill, EINVAL is returned in
most cases.

Change those cases to ESRCH, which is what will be returned if
the code makes it to the kill(2) call and the process has
disappeared at that point.

Add a unit test.