broker: suppress runat kill ESRCH errors
Problem: when a flux instance is signaled during shutdown, brokers may
log "rc3: error aborting" if the signal is not delivered to a running
rc3 process, but these message are not helpful and are confusing to users.

Suppress that message if the error is ESRCH.  Also change the message from
"rc3: error aborting: <strerror>" to "kill rc3: <strerror> to make it more
descriptive.

Note that the message can be logged from two different points in runat.c.
Both are updated.