shell: input: add redirect event with input from file
Problem: Now that file input bypasses the KVS, no input eventlog is
created in this case. Previously, an input eventlog was always created,
so its absence now might cause issues or confusion.

When redirecting input from a file, create the input eventlog with
a header and a RFC 24 "redirect" event with the path to the input file.