shell: do not fail on oom init failure
Problem: The oom detector in the job shell is not critical for job
execution.  However, if it fails to be initialized, the job shell
will exit with a shll.init failure and the job will fail.

Solution: Output a warning that oom detection is disabled when initialization
fails.  Do not fail to execute the job-shell.  The user should now see
something like the following for an inotify error.

2.636s: flux-shell[113]:  WARN: oom: disabling oom detection: max number of user inotify instances has been reached

Fixes #5920