broker: configure child recv queue limit
Problem: the memory resident set size has been observed to get
out of hand on the leader node of a system that has has a flat
TBON 1:N with N in the thousands.

The zmq message queues are thought to be part of the problem.
Currently the child socket, which fans out to all the downstream
peers, allows an unlimited number of messages to be queued, and
when the system is busy, that queue can grow arbitrarily large.
Allow ZMQ_RCVHWM to be configured on that socket so the number
of queued messages per peer can be limited.

Experiments with libzmq-4.2.1 show when the limit is reached, messages
queue on the sender side without blocking zmq_send(), and when space is
available again, message forwarding resumes.  So setting a limit may be
advantageous when the fanout is large and the system is busy, since memory
utilization for messages is distributed across children rather than
concentrated on their parent.

Add a broker attribute and a config key so we can experiment with these
settings.  The default is still 0 (unlimited).