libsubprocess: remove remote input prep/check
Problem: Profiling shows that a significant amount of time
can be spent in the prep/check of remote subprocess input.
This is even in the case when the input buffer is empty.

It ends up that the prep/check is not necessary for remote input.
If the subprocess is already running, it can be written to directly
without buffering.  Buffering is only needed when a caller attempts
to write to the subprocess before the subprocess is running.

For remote subprocesses, remove all channel input prep/check.  Immediately
write to the remote subprocess if the subprocess is running.  If the subprocess
is not yet running, buffer the input and write it out later.