job-exec: retry SIGKILL of job shell while job is still running
Problem: The job-exec module only attempts to send SIGKILL once to
job shells after the job shell kill timeout has elapsed. If for some
reason the signal is not delivered to one or more shells, then the
job can hang forever in CLENAUP state waiting for thos shells to
exit.

Retry the send of SIGKILL to job shells as long as the job is still
active in the exec system. Since kernel or filesystem issues could
cause a shell to become unkillable, implement a backoff scheme so
that the kill is not sent every 5s in this case.

Fixes #5967