libsubprocess: create write_buffer on demand
Problem: Profiling shows that creating the write buffer for
remote subprocesses eats up a healthy amount of cycles.  However,
the buffer is not needed for many circumstances.  It is only needed
when there is an attempt to write data to the subprocess before
the subprocess is running.

Create the write buffer only when it is needed.