sdexec: handle early sdexec.write requests
Problem: if an sdexec.write request arrives before the unit stdin
is ready, it will be dropped.

This is possible now that sdexec.write identifies the unit by the
matchtag instead of the pid.

Queue early sdexec.write requests until stdin is valid, then move
them back to the flux handle for processing as though they are being
received for the first time.