job-exec: add kill-signal and term-signal parameters
Problem: It would be convenient for testing to redefine the default
signals that the job-exec module uses for task and job shell
termination for testing purposes, but this currently is not supported.

Add support for setting `term-signal` (default SIGTERM) and kill-signal
(default SIGKILL) in configuration and the module command line.