job-exec: rename and improve job_exec_initialize()
Problem: The job_exec_initialize() function is in reality just a
specialized config parsing function that handles configuration of
some internal global parameters in the module.

Rename the function job_exec_set_global_config() for clarity. Return
errors in a flux_error_t instead of logging them so that this function
could be used in the reconfig callback in the future.