libsubprocess: respond with line buffered output
Problem: If line buffering is enabled on a remote subprocess, the
server does not respond with output containing a single line of
data.  This was not an issue when output data was buffered on the
client side, but with the new LOCAL_UNBUF flag, multiple lines of
data could be read in a single callback.

In the server, get flux_subprocess_getline() over flux_subprocess_read().
Only use flux_subprocess_read() if line buffering was not enabled.