testsuite: expand job-exec config testing
Problem: There are new exec config parameters that are not tested for
correct handling in the testsuite. Also, some of the config globals
in job-exec: kill-timeout, kill-signal, and term-signal, can now be
updated via a config reload, and this is also untested.

Expand testing in t2403-job-exec-conf.t to cover these new
capabilities.