libsdexec: avoid double EOF notification
Problem: sdexec sends two EOF notifications for each channel, which
causes libsubprocess to get confused and not signal subprocess
completion when the LOCAL_UNBUF flag is used.

The sdexec channel class stops its own watcher after sending EOF,
but the sdexec module restarts the watcher during completion to ensure
it can send EOF in an error case where it was not started.  This
results in two EOF notifications.

Add an 'eof_delivered' flag to the channel and use it to ensure that
only one eof is sent.

Fixes #6019