job-exec: track number of attempts to signal job
Problem: It may be useful to determine how many attempts to kill the
job and/or job shell has been made by the job exec system, but this
information is not tracked.

Add kill_count and kill_shell_count to struct jobinfo and
increment each after a successful call to flux_job_kill(3) or
flux_subprocess_kill(3) respectively.