job-exec: add running job information to module stats
Problem: When there's a problem with one or more running jobs it
would be useful to get some information from the job-exec module,
but there's currently no way to do that.

Add internal information about running jobs to the job-exec module
stats object. Each job is placed in a "jobs" dictionary, keyed by its
jobid, with contents including much of the internal state for that job.

For example:

 "jobs": {
  "ƒUabC7Dy": {
   "implementation": "bulk-exec",
   "ns": "job-1049901400064",
   "critical_ranks": "0",
   "multiuser": 0,
   "has_namespace": 1,
   "exception_in_progress": 1,
   "started": 1,
   "running": 1,
   "finalizing": 0,
   "kill_timeout": 5.0,
   "kill_count": 0,
   "kill_shell_count": 0,
   "total_shells": 1,
   "active_shells": 1
  }