job-exec: improve exception message for lost shell
Problem: The exception logged when the job-exec module loses contact
to a job shell can be confusing to users, who may not know what a
job shell is and may conclude the problem is a Flux failure and not
a node failure, which is typically the case.

Change the error message from "lost contact with job shell on <node>"
to "node failure on <node>" to make the message more clear.

Update one test that was checking for "lost shell" in an error message.