libsubprocess: use matchtag in client write req
Problem: RFC 42 now requires write requests to reference the
subprocess via the exec matchtag rather than the pid.

Send the matchtag in the write request instead of the pid and
alter the internal function prototype for subprocess_write() to
accept the exec future in lieu of rank, service, and matchtag.
Update users of that function.

Also update a unit test that uses the raw protocol.