sdexec: fix segfault during teardown
Problem: the sdexec module can segfault when a subprocess unit
is destroyed as a result of an output channel closing.

The channel output callback destroys the channel, then the channel
code continues referencing it after the callback.  Add a reference
count that is held over the callback and the duration of hte function
to prevent the segfault.

Fixes #6036