completions: improve and enhance option handling
Problem: Long options are not handled well in the flux bash
completions, making it difficult to generation completions when the
long option does vs does not include the '=' (.e.g --config-file=
vs --config-file), or when a short option equivalent is used instead
(e.g. `-c`).

Adopt the _split_longopt() function from the bash-completion
project, which splits long options like `--foo=bar` into `cur=bar`
and `prev=--foo` so that completions can easily be supported for
`--foo<space>` or `--foo=`.  This also makes it possible to support
single option equivalent `-f bar` in the same manner.

This simplifies the current _flux_complete_queue() and
_flux_complete_format_name() which now don't need to check $cur and
$prev for matching patterns, but can be functions that simply return
the current list of queues or format names.

Update all current completions where it makes sense with a case
statement for handling special case options (such as `-q, --queue=`
or `-o, --format=`).  Add additional handlers where possible to return
sets of completions for specific options.

Short options that require an argument and for which there are no
suggestions are handled in the case statement with an immediate
return.  This offers no completions so that a tab doesn't suggest
other command options in place of a reqiured argument. (Handling long
options similarly is optional since functions return immediately if
a long option is split.)

Other fixes and improvements are lumped into this update as they were
found during testing of the conversion.