systemd: add prolog/epilog service units
Problem: prolog/epilog scripts should be run in a systemd cgroup
for reliable termination, logging to the local systemd journal, and
debugging using well known systemd tools.

Add a systemd "oneshot" service unit for prolog and epilog, templated by
jobid.  The unit runs a user-provided /etc/flux/system/prolog or epilog
script.  The unit is configured so that
- the environment can be passed in via a file in /run
- 'systemctl start' blocks the the run is complete
- the 'systemctl start' exit code reflects the exit code of the script
- the unit instance is automatically garbage collected (even on failure).

Add corresponding helper scripts that can be configured as an IMP run
command.  The unit name template is filled in wtih the job id from the
FLUX_JOB_ID environment variable, then
- flux-run-prolog runs 'systemctl start flux-prolog@<jobid>'
- flux-run-epilog runs 'systemctl start flux-epilog@<jobid>'

The environment of the helper script is copied to the unit script via
a temporary file created in run.

The helper script handles SIGTERM by running 'systemctl stop'.  This
should be a fairly reliable termination method (as reliable as possible
from user space anyway), even if the helper script is subsequently killed
with SIGKILL.