docker: switch to podman for 'system' test container
Problem: On newer distros, the container in docker-run-systest.sh
fails to run with an error like

 Failed to mount tmpfs at /run: Operation not permitted

It seems docker cannot run with sufficient privliege to setup the
correct tmpfs based filesystems necessary for systemd.

Switch docker-run-systest.sh to use podman with the following options
that appear to support systemd in a container

 podman --systemd=always \
        --userns=keep-id \
        --security-opt unmask=/sys/fs/cgroup

Due to an issue with the v3.4.x version of podman available in the
GitHub actions container, use `docker save` and `podman load` to make
the container built by docker available to podman.

Additionally, podman saves the image as the wrong name in this version,
so use the container id (digest) instead.