shell: add flux_shell_get_hostlist(3)
Problem: Getting the list of hostnames assigned to the job is a common
requirement for job shell plugins, but the shell API doesn't provide
a function to do it.

Add flux_shell_get_hostlist(3) which obtains the assigned hostlist
from R, caches it, and returns the result to the caller.