shell: doom: report hostname in log messages
Problem: The doom plugin doesn't report hostnames on which the tasks
that exited early were running, but this would be very useful to
users and admins when the early exiting tasks were not expected.

Report hostnames for affected tasks (or job shell) in the error
message when raising an exception in the doom plugin.