job-manager/prioritize: update priority _before_ sending
problem: On receiving a message from `flux job urgency` the job-manager
prioritize code updates the value of job->priority after sending the
update to the scheduler, using the value of job->priority.  As a result,
the scheduler receives an update RPC setting the priority to whatever it
was last set to and only on the next such update does it receive the
requested value.

solution: update job->priority before sending the RPC to the scheduler