actions: only run CI for push on master
problem: we currently run CI in full on both push and PR for all
branches, and it mostly just fails on push because it doesn't receive
label context. What we really need is for push triggered builds to run
on master and only PR-triggered otherwise.

solution: limit push event triggers to the master branch.