man: create include file for EXPERIMENTAL
Problem: the text to explain expectations for experimental features
must be repeated in documentation for each one which is extra work
for both the author and the reader.

Add common/experimental.rst which can be included from man pages
that discuss experimental features.  Like common/resource.rst, the
idea is to make it a standalone section which can then be referenced
from the feature documentation.

This should improve consistency and make it easier to document these
features and interfaces.

Fixes #6066