libsubprocess: add new accessors to bulk-exec API
Problem: It would be useful to get the number of completed processes
and the active ranks from the bulk-exec interface, but there are no
functions currently exported to get this data.

Add bulk_exec_complete() to get the count of completed processes for
the current bulk-exec object, and bulk_exec_active_ranks() to get the
ranks on which procesess are still active.