job-exec: improve bulk exec per-job stats
Problem: The current job-exec per-job stats use the wrong bulk-exec
function to get the count of active shells in the job. Also, it would
be useful to report the ranks of active shells when a job is stuck on
a subset of ranks.

Use `total - complete` to calculate `active_shells` in the job-exec
per-job stats. Additionally, add a new `active_ranks` member with the
idset of ranks on which job-exec thinks job shells are active to the
job-exec per-job stats.

Fixes #6068